requirements.txt
discord.py
python-dateutil
requests
pyenchant
beautifulsoup4
PySocks
lxml
nltk
matplotlib
wolframalpha
Pillow
APNGLib
DerPyBooru
imageio
dice
PyNaCl
youtube_dl
